<Feat>:Add local h5->xmodel files
Adding the files i use in order to create an xmodel from .h5.
Here, i include the arch,json with improved clockspeed, an .h5 model that has over 90% acc on inference, and an .ipynb that compiles the .h5 to .xmodel so that it runs on the Kria with more than 90% acc on the fpga.